requirements.txt
Django==1.9.1
django-filter==0.12.0
djangorestframework==3.3.2
Markdown==2.6.5
psycopg2==2.6.1
